To, The Headmaster PM SHRI MPS Varsha Nagar Vikhroli West, Mumbai-79
Date: _____________

Subject: Request for Late Arrival / Early Departure

Respected Sir/Madam,
I, _____________________ (Student's Name), of Grade _____, Division _____, request permission to:
☐ Arrive late at ______ AM/PM
☐ Leave early at ______ AM/PM
on _____________ (Date).

Reason: _______________________________________

Kindly allow accordingly.
Thank you.
Yours faithfully,
(Signature)    it should be in this format and make a button in it as download form in pdf when he fill the form and and send to approvel and same thing should be done for the teacher also teacher can download the form when form is approved